+++ packages/spfx/src/getPageLanguage/index.ts
@@ -1,0 +1,9 @@
+import { getString } from "@datyin/core";
+import { getContext } from "../index";
+
+function getPageLanguage() {
+  const context = getContext();
+  return getString(context?.pageContext?.cultureInfo.currentUICultureName).toLowerCase();
+}
+
+export { getPageLanguage };