+++ packages/spfx/src/getPageName/index.ts
@@ -1,0 +1,9 @@
+import { getString } from "@datyin/core";
+import { getContext } from "../index";
+
+function getPageName(): string {
+  const context = getContext();
+  return getString(context?.pageContext?.web?.title);
+}
+
+export { getPageName };