+++ packages/spfx/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+`2023-07-09` - `0.0.3`
+- Added `getPageName` function
+- Added `getPageLanguage` function
+
 `2023-07-09` - `0.0.2`
 - Fixed missing email address on `getCurrentUser`